requirements.txt
python-dotenv==1.1.1
pyodbc==5.3.0
psycopg2-binary==2.9.11
fastapi==0.119.1
uvicorn==0.38.0
pyyaml==6.0.3
cryptography==46.0.3